Set max numpy version in pyproject.toml
This change is intended to avoid encountering issue #13 ("numpy.ufunc
has the wrong size") by ensuring that the version of numpy that the
gtar module was compiled against is lower or equal to the version that
will be used at runtime. Users that know what they are doing can build
the gtar module using the numpy that is already installed with `pip
install --no-deps --no-build-isolation` or other similar commands.